Fixed Login Exception by throwing it
You can decide if you want to handle this by an exception or choose another strategy. If so, remove the exception.  Or keep it up to you ;)

Good Work with the navigation and general design of course ( 😊 ) , but we need to pay attention to all aspects of the project. If we have something in our code, it should be there for a reason. We don’t want dead code to be laying around (Exceptions that is not thrown but marking that a method throws it.